compiler: bugfix: rest parameter in inline lambda
* share/txr/stdlib/compiler.tl (lambda-apply-transform): Do
not take all of the fixed arguments and rest expression to be
the trailing list. Rather, skip as many elements from these
as the function has fixed parameters. E.g. if there
are two fixed parameters as in (lambda (a b . c)) and the
call specifies four fixed parameters and a trailing x
(1 2 3 4 . x) then the rest argument c must be (list* 3 4 . x)
and not (list* 1 2 3 4 . x).