compiler: bug: graph issue in dead code elimination.
* share/txr/stdlib/optimize.tl (basic-blocks join-block): When
we merge a block N with its physical predecessor P, we must
iterate over the jump targets of N, and update their reverse
list to point to P instead of N. A full call to link-graph
would fix it also, but we don't call link-graph anywhere in
the optimization pipeline after this point. The reverse
links are relied on to be correct elsewhere, such
merge-jump-tunks, which needs to walk the blocks which jump
to a removed duplicate block, to retarget them to go to
another copy of that block.