sha256/md5: leak: don't use borrowed buffer for hash.
It's incorrect to use a borrowed buf, because a borrowed buf
assumes that the memory isn't dynamically allocated, and will
not free it when it is finalized by the garbage collector.
The buffer which holds the hash returned by sha256 and md5
does in fact own the memory.

* buf.c (make_owned_buf): Static function becomes external.

* buf.h (make_owned_buf): Declared.

* chksum.c (chksum_ensure_buf): Use make_owned_buf rather than
make_borrowed_buf.