listener: don't complete on unbound symbols
This patch prevents Tab-completing on interned symbols that
have no binding.

The current behavior is, for instance:

  1> 'hamsandwich
  hamsandwich
  2> 'ham[Tab]
  2> 'hamsandwich ;; completes

The new behavior will not complete hamsandwich, because it has
no binding as a function or variable.

* parser.c (find_matching_syms): Treat the default case the
same as after '[': a function or variable binding is required,
or the symbol is not listed. Use fboundp instead of
lookup_fun.  They are the same, except in TXR 127 compat mode,
which includes macros under fboundp.