compiler: bugfix: warnings deferred too far.
Because with-compilation-unit is keyed of *load-recursive*,
when compilation is happening in the context of a load (a
top-level form in a loaded file calls compile-file or
compile-update file) warnings are deferred until the end of
the load. That might never occur if the load doesn't complete,
because, say, the image quits for some reason.

If the following is the content of a file which is loaded:

  (compile-file "foo")
  (exit 0)

then warnings during the compilation are not issued when
compile-file terminates, and will never be issued because of
the termination due to the exit call.

* share/txr/stdlib/compiler.tl (*in-compilation-unit*): New
special variable.
(with-compilation-unit): Use *in-compilation-unit* to
determinw when a compilation unit has ended, and dump all the
deferred warnings then. We will bind *load-recursive* because
that is required for deferring warnings.