json: implement distinguished json quasiquote.
Because #J<json> produces the (json ...) form that translates
into quote, ^#J<json> yields a quasiquote around a quote. This
has some disadvantages, because it requires an explicit eval
in some situtions to do what the programmer wants.

Here, we introduce an alternative: the syntax #J^<json> will
produce a quasiquote instead of a quote.

The new translation scheme is

  #J X  -> (json quote <X>)

  #J^ X -> (json sys:qquote <X>)

where <X> denotes the Lisp object translation of JSON syntax X.

* parser.c (me_json): The quote symbol is now already in the
json form, so all that is left to do here is to take the cdr
to pop off the json symbol.

* parser.l (JPUNC, NJPUNC): Allow ^ to be a punctuator in
JSON mode.

* parser.y (json): For regular #J, generate the new (json
quote ...) syntax. Implement J# ^ which sets up the nonzero
quasi_level around the processing of the JSON syntax, so that
everything is in a quasiquote, finally producing the
(json sys:qquote ...) syntax.

* lex.yy.c.shipped, y.tab.c.shipped: Updated.