json: functions put-json and put-jsonl.
* eval.c (eval_init): Register put-json and put-jsonl
intrinsics.

* lib.c (out_json_str): Do not output the U+DC01 to U+DCFF
code points by masking them and using put_byte. This is
unnecessary; if we just send them as-is to the text stream,
the UTF-8 encoder does that for us.
(put_json, put_jsonl): New functions.

* lib.h (put_json, put_jsonl): Declared.

* txr.1: Documented. The bulk of tojson is moved under the
descriptions of these new functions, and elsewhere where the
document pointed to tojson for more information, it now points
to put-json. More detailed description of character treatment
is given.

* share/txr/stdlib/doc-syms.tl: Updated.