parser: gc bug in token.
The parser maintains some token objects for one parse job to
the next in the tok_pushback and recent_tok arrays.  When the
recent_tok is assigned into the parser, it could be a
wrong-way assignment: the parser is a gen 1 object,
yet the token's semantic value of type val is a gen 0.

* parser.c (lisp_parse_impl, txr_parse): Before re-enabling
gc, indicate that the parser object may have been mutated by a
wrong-way assignment using the mut macro.

* txr.c (txr_main): Likewise.