json: fix quasiquote print-read consistency issue
* lib.c (out_json_rec): When printing keys that might be
potentially quasiquoted symbols that look like ~abc, we
must avoid the condensed {~abc:~def} style. This is because
abd:~def looks like a single symbol, where abc is a package
qualifier. To be safe and readable at the same time, we add
spaces whenever either the key or the value are conses,
indicating non-JSON syntax. Spaces are added on both sides of
the colon, and also after the preceding comma, if there
is a previous item.