unwind: unbind dyn env when diagnosing unhandled exception.
* unwind.c (uw_unwind_to_exit_point): The error reporting
relies on the *stderr* stream, but that stream may be bound in
some funny way in a context where an exception has happened,
without anything around it to restore that during unwinding.
Before diagnosing, we should reset the dynamic environment to
the top level, so that the global values of all dynamic
variables are revealed.